% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmCorrs.R
\name{corstars}
\alias{corstars}
\title{Corstars}
\usage{
corstars(
  x,
  group = NULL,
  method = "pearson",
  removeTriangle = c("upper", "lower"),
  alpha.order = FALSE,
  stars = 2,
  result = "html",
  sumstats = TRUE,
  title = "Correlation Table"
)
}
\arguments{
\item{x}{Data object.}

\item{group}{Optional grouping variable as a quoted string (e.g., group = "sex").}

\item{method}{Correlation method. Default is pearson.}

\item{removeTriangle}{Default is upper (per APA).}

\item{alpha.order}{Alphabetize variables. Default is FALSE.}

\item{stars}{Number of significance stars. Default is 2, max is 4 (p < .0001).}

\item{result}{Output options. Default is "html". Option "text" returns a data frame.}

\item{sumstats}{Include mean, SD, and N. Default is TRUE.}

\item{title}{Table caption.}
}
\value{
A correlation table (gt object, data frame, or xtable)
}
\description{
This function creates a correlation matrix with descriptive statistics.
}
