% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmCorrs.R
\name{corstars}
\alias{corstars}
\title{Corstars}
\usage{
corstars(x, method = "pearson", removeTriangle = c("upper", "lower"),
  alpha.order = F, result = "html", sumstats = T,
  title = "Correlation Table")
}
\arguments{
\item{x}{Data object.}

\item{method}{Correlation method. Default is pearson}

\item{removeTriangle}{Default is upper (per APA).}

\item{alpha.order}{Alphabetize variables.  Default is FALSE.}

\item{result}{Output options.  Default is html table to viewer.  Option "text" returns output to console only.}

\item{title}{Table caption.}
}
\value{
A correlation table
}
\description{
This function creates the ICC matrix
}
