% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmCorrs.R
\name{icc.corrs}
\alias{icc.corrs}
\title{Estimate ICCs, correlations, and descriptive Statistics}
\usage{
icc.corrs(x, group, title = "Descriptive Stats", gmc = FALSE,
  alpha.order = FALSE, result = "html")
}
\arguments{
\item{x}{Data object.}

\item{group}{Nesting variable.}

\item{title}{Table caption.}

\item{gmc}{Provide group-mean center correlations.  Default is FALSE.}

\item{alpha.order}{Sort variables alphabetically.  Default is False}

\item{result}{Output options.  Default is html table to viewer.  Option "text" returns output to console only.}
}
\value{
A correlation table with sample statistics and ICC estimates
}
\description{
This function creates the ICC matrix
}
