% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmCorrs.R
\name{lgm}
\alias{lgm}
\title{Latent Group Model}
\usage{
lgm(x, group, title = "LGM", stars = TRUE, sumstats = TRUE,
  sumtable = FALSE, alpha.order = FALSE)
}
\arguments{
\item{x}{Data object.}

\item{group}{Nesting variable.}

\item{title}{Table capation.}

\item{stars}{Provide significance stars.  Default is TRUE.}

\item{sumstats}{Provide summary stats.  Default is TRUE.}

\item{sumtable}{Provide sem summary table. Default is FALSE.}

\item{alpha.order}{Sort variables alphabetically.  Default is FALSE.}
}
\value{
A correlation table
}
\description{
This function creates the LGM matrix
}
